pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qmt-quant-demo"
version = "0.1.0"
description = "基于 QMT 的事件驱动量化交易 Demo 平台"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Codex" }]
dependencies = [
  "pandas>=2.0",
  "pydantic>=2.0",
  "PyYAML>=6.0",
  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.3",
]

[project.optional-dependencies]
ui = ["streamlit>=1.0"]
api = ["fastapi>=0.115", "uvicorn>=0.30"]
test = ["pytest>=8.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q --basetemp=data/pytest_tmp"
